+++ Bresenham-Line/Seconn/OutputLayer.swift
@@ -9,15 +9,15 @@ import Foundation
 import Surge
 
 struct OutputLayer {
-    let reductionMatrix: SliceableMatrix<FloatType>
-    let reductionMatrixTransposed: SliceableMatrix<FloatType>
+    let reductionMatrix: SliceableMatrix<Double>
+    let reductionMatrixTransposed: SliceableMatrix<Double>
 
     init(inputSize: Int, outputSize: Int) {
         precondition(inputSize % outputSize == 0, "Output layer should be a multiple of its input")
         let count = inputSize / outputSize
-        let rate = FloatType(outputSize) / FloatType(inputSize)
-        var matrix = SliceableMatrix(rows: outputSize, columns: inputSize, repeatedValue: FloatType(0.0))
-        var rowArray = Array(repeating: FloatType(0.0), count: inputSize)
+        let rate = Double(outputSize) / Double(inputSize)
+        var matrix = SliceableMatrix(rows: outputSize, columns: inputSize, repeatedValue: Double(0.0))
+        var rowArray = Array(repeating: Double(0.0), count: inputSize)
         for i in 0..<outputSize {
             let prefixCount = count * i
             let postfixCount = count * (outputSize - i - 1)
@@ -48,14 +48,14 @@ extension OutputLayer: Layer {
         return reductionMatrix.rowCount
     }
 
-    func process(input: [FloatType]) -> [FloatType] {
+    func process(input: [Double]) -> [Double] {
         return Array((reductionMatrix * SliceableMatrix(column: input))[column: 0])
     }
 }
 
 
 extension OutputLayer: LearningLayer {
-    mutating func learn(input: [FloatType], output: [FloatType], target: [FloatType], weightRate: FloatType, biasRate: FloatType) -> [FloatType] {
+    mutating func learn(input: [Double], output: [Double], target: [Double], weightRate: Double, biasRate: Double) -> [Double] {
         return ceil(clip((reductionMatrixTransposed * SliceableMatrix(column: target))[column: 0], low: 0.0, high: 1.0))
     }
 }
